Redefine laptops table

+++ database/migrations/2019_06_13_085516_create_laptops_table.php
@@ -17,7 +17,7 @@ class CreateLaptopsTable extends Migration
             $table->bigIncrements('id');
             $table->string('brand');
             $table->string('name');
-            $table->decimal('memory');
+            $table->integer('memory');
             $table->decimal('price');
             $table->timestamps();
         });
